Recognize hyphenated names and names with apostrophes
This code now works with hyphenated names and names with apostrophes, as
long as they begin with either an 'M' or an 'R'. Ashley and I agreed via
email that this is the correct rule.

There are still issues with POS tagging: names with apostrophes do not
necessarily get tagged as PROPN. That means that the patterns do not
necessarily match, which will create problems for pattern recognition
later.